fix(barcode): do not stringify array-valued custom barcode meta
A custom barcode meta key can hold anything another plugin put there.
The new (string) cast turned an array into the literal "Array" plus a PHP
warning, where the old trait returned get_meta() raw. Only scalars are a
barcode now; anything else reads as ''.

Also drops the essential-meta allowlist's now-unreachable ! empty() check
(meta_key() never returns '') and states why, and moves the test double
into its own file so the test dir stays PHPCS-clean if the ruleset ever
covers tests/.
